comprehensive test

+++ src/main/java/org/jointheleague/ventilator/stepper/StepperInterface.java
@@ -1,7 +1,7 @@
 package org.jointheleague.ventilator.stepper;
 
 public interface StepperInterface {
-	void forward(int rate, int time);
+	void forward(double rate, double time);
 	
-	void backward(int rate, int time);
+	void backward(double rate, double time);
 }
